// src/main/users/interfaces/users.ts
import { Document } from 'mongoose';

export interface Users extends Document {
  email: string;
  password: string;
  localToken: string;
  passwordResetToken: string;
  passwordResetExpires: Date;
  facebook: string;
  tokens: string;
  userId: string;
  attorneyName: string;
  attorneyId: string;
  userName: string;
  departamentId: string;
  legalName: string;
  userRole: any;
  legalFormId: string;
  legalActivityId: string;
  companyTaxId: string;
  entity: number;
  googleAddress: {
    country: string;
    region: string;
    province: string;
    city: string;
    postalCode: string;
    street: string;
    streetNumber: string;
    postfixAddress: string;
    longitude: number;
    latitude: number;
    formattedAddress: any;
  };
  postalCode: string;
  location: string;
  street: string;
  streetNumber: string;
  postfixAddress: string;
  longitude: string;
  latitude: string;
  telephone: string;
  fax: string;
  gdprId: string;
  quoteId: string;
  policiesId: string;
  Bank1: string;
  AccountHolder1: string;
  Account1: string;
  Clabe1: string;
  Bank2: string;
  AccountHolder2: string;
  Account2: string;
  Clabe2: string;
  subAccount: string;
  balanceId: string;
  commentsText: string;
  active: string;
  recordLimit: number;
  kardexId: string;
  logo: string;
  perfilImg: string;
  comparePassword: string;
  gravatar: (size: number) => string;
  updatePassword: boolean;
  wildduckEmail: string;
  wildduckPassword: string;
  emailSignature: string;
  createdBy: any;
  updatedBy: any;
}